Slow down QueueConsumer to improve batching
Batching is handled by the NonBlockingStatsDClient's QueueConsumer. It
consumes messages from a queue, places them in a buffer, and then
flushes if the queue had been drained or if the maximum batch size has
been reached.

As the QueueConsumer does a small amount of work per message, it's a
fast consumer -- and since it flushes right as the queue is drained, it
tends to dispatch a batch of size 1.

This commit attempts to improve batching by slowing down the consumer.
If we drain the queue, but have recently flushed, hold back and
optimistically wait for more messages.